+++ tests/testthat/test-tabraux.R
@@ -1,4 +1,4 @@
-
+library(tabr)
 ## tests for expand_notes
 i9 = 'a3  b r c4 g3 f'
 o9 = "a3 b3 r c#4 g3 f#3"
@@ -35,3 +35,31 @@ test_that("check_times", {
   expect_identical(o16$counts,rep(6,4) )
 })
 ## tests for lilypond_version
+o17 = lilypond_version()
+test_that("lilypond_version", {
+  expect_identical(class(o17),"character" )
+  expect_true(grepl('^[.0123456789]{7,8}$',o17))
+})
+## tests for edit_phrase
+notes1 <- "e f g f"
+info1 <- "4*4"
+phrase1 <- p(notes1,info1)
+track1 <- track(phrase1,key = "a",voice = 1,tab=F)
+o18 = edit_phrase(phrase1,'f','a',all=T )
+o19 = edit_phrase(phrase1,'f','a',all=F )
+o20 = edit_phrase(track1,'f','a',all=T )
+o21 = edit_phrase(track1,'f','a',all=F )
+test_that("edit_phrase", {
+  expect_error(edit_phrase('my char','char','CHAR',all=T ),"edit_phrase: invalid type for obj")
+  expect_identical(class(o18),c("phrase","character") )
+  expect_identical(as.character(o18),"<e>4 <a>4 <g>4 <a>4")
+  expect_identical(class(o19),c("phrase","character") )
+  expect_identical(as.character(o19),"<e>4 <a>4 <g>4 <f>4")
+  expect_identical(class(o20),c("track", "tbl_df", "tbl", "data.frame") )
+  expect_identical(class(o20$phrase[[1]]),c("phrase","character") )
+  expect_identical(as.character(o20$phrase[[1]]),"<e>4 <a>4 <g>4 <a>4")
+  expect_identical(class(o21),c("track", "tbl_df", "tbl", "data.frame") )
+  expect_identical(class(o21$phrase[[1]]),c("phrase","character") )
+  expect_identical(as.character(o21$phrase[[1]]),"<e>4 <a>4 <g>4 <f>4")
+})
+