+++ NEWS.md
@@ -1,6 +1,7 @@
 ### tabraux 0.0.2 01jan2019
 
 - Function `lilypond_version` is added
+- Function `edit_phrase` is added
 - Function `pc2` is removed. `tabr::pc` has now the same functionality
 
 ### tabraux 0.0.1 20dec2019